+++ src/constants/common.js
@@ -1,0 +1,3 @@
+export const HOME = -1;
+export const COMPLETED = 0;
+export const REMAINING = 1;
