+++ src/utils/utils.js
@@ -1,7 +1,3 @@
-export const HOME = -1;
-export const COMPLETED = 0;
-export const REMAINING = 1;
-
 export const getLocalStorageItem = (item) =>
   window.localStorage.getItem(item);
